requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
Flask>=3.0.0
gunicorn>=21.0.0
